Alot of stuff
Okay so I did some training and played around with a few ideas. Finally
settled on hard coding specific moves at the start of the game and that
seems to be working quite well. I want to explore that idea further
since it should be possible to code out the ideal game start up to a
point. I also want to do some more training with the new starts and
building on previously trained players to strengthen the mid game.
Hoping for the best. Not really sure how much more I can do about this